@@ -1,3 +1,4 @@
+using ApiTaCerto.Models.Usuario;
 using System.ComponentModel.DataAnnotations;
 
 namespace ApiTaCerto.Models
@@ -7,10 +8,14 @@ namespace ApiTaCerto.Models
         [Key]
         public int IdAtividadeAluno { get; set; }
         public int NumeroTentativas { get; set; }
+        public int IdPessoa { get; set; }
+        public int IdAtividade { get; set; }
         public double MaiorNota { get; set; }
         public int MenorTempo { get; set; }
         public int MaiorTempo { get; set; }
-        public int IdPessoa { get; set; }
-        public int IdAtividade { get; set; }
+
+        //NAVIGATION PROPERTY
+        public Pessoa Pessoa { get; set; }
+        public Atividade Atividade { get; set; }
     }
 }
