@@ -1,10 +1,16 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using ApiTaCerto.Models.Usuario;
 using Newtonsoft.Json;
 
 namespace ApiTaCerto.Models
 {
+    public enum Origem
+    {
+        TaCerto,
+        TaCertoForms
+    }
     [JsonObject, Serializable]
     public class LogLogin
     {
@@ -18,6 +24,9 @@ namespace ApiTaCerto.Models
         public string DeviceId { get; set; }
         [MaxLength(150)]
         public string DeviceIp { get; set; }
-        public int Origem { get; set; }
+        public Origem Origem { get; set; }
+
+        // NAVIGATION PROPERTY
+        public Pessoa Pessoa { get; set; }
     }
 }
