@@ -18,7 +18,7 @@ namespace ApiTaCerto.Repositorio
 
         public Disciplina Find(long id)
         {
-            return _contexto.Disciplinas.FirstOrDefault(p => p.IdDisciplina == id);
+            return _contexto.Disciplina.FirstOrDefault(p => p.IdDisciplina == id);
         }
 
         public IEnumerable<Disciplina> GetAll()
@@ -28,12 +28,16 @@ namespace ApiTaCerto.Repositorio
 
         public IEnumerable<Disciplina> GetAllSubjectsWithId(int id)
         {
-            return _contexto.Disciplinas.Where(d => d.IdMatriz == id).ToList();
+            return _contexto.Disciplina.Where(d => d.IdMatriz == id).ToList();
         }
 
         public int GetDefaultInstituteId(string cnpj)
         {
-            return _contexto.Instituicoes.Where(i => i.CNPJ == cnpj).First().IdInstituicao;
+            Instituicao instituicao = _contexto.Instituicao.Where(i => i.CNPJ == cnpj).FirstOrDefault();
+            if (instituicao != null)
+                return instituicao.IdInstituicao;
+
+            return 0;
         }
 
         public IEnumerable<AtividadeDisciplina> GetAllClassActivities(long idDisciplina){ 
@@ -41,22 +45,22 @@ namespace ApiTaCerto.Repositorio
         }
 
         public IEnumerable<Atividade> GetAllActivities(long idDisciplinaAutor){
-            return _contexto.Atividades.Where(a => a.IdTurmaDisciplinaAutor == idDisciplinaAutor && a.NumeroQuestoes > 0).ToList();
+            return _contexto.Atividade.Where(a => a.IdTurmaDisciplinaAutor == idDisciplinaAutor && a.NumeroQuestoes > 0).ToList();
         }
 
         public IEnumerable<Questao> GetAllQuestions(long idAtividade){
-            return _contexto.Questoes.Where(a => a.IdAtividade == idAtividade).ToList();
+            return _contexto.Questao.Where(a => a.IdAtividade == idAtividade).ToList();
         }
 
         public int GetAttemptsNumber(long idAtividade){
-            int maxAttempts = _contexto.Atividades.Where(a => a.IdAtividade == idAtividade).FirstOrDefault().NumeroTentativas;
+            int maxAttempts = _contexto.Atividade.Where(a => a.IdAtividade == idAtividade).FirstOrDefault().NumeroTentativas;
 
             return maxAttempts;
         }
 
         public IEnumerable<Disciplina> GetAllClassSubjects(long idTurma){
             using(var contexto = _contexto){
-                var innerJoin = from disc in contexto.Disciplinas join disctur in contexto.DisciplinaTurmas on disc.IdDisciplina equals 
+                var innerJoin = from disc in contexto.Disciplina join disctur in contexto.DisciplinaTurma on disc.IdDisciplina equals 
                 disctur.IdDisciplina where disctur.IdTurma == idTurma select disc;
 
                 List<Disciplina> atividades = innerJoin.ToList();
@@ -67,36 +71,36 @@ namespace ApiTaCerto.Repositorio
         public async Task<int> SaveAtividadeResposta(AtividadeRespostaAluno atividadeRespostaAluno)
         {
 
-            await _contexto.AtividadeRespostaAlunos.AddAsync(atividadeRespostaAluno);
+            await _contexto.AtividadeRespostaAluno.AddAsync(atividadeRespostaAluno);
             await _contexto.SaveChangesAsync();
 
             return atividadeRespostaAluno.IdAtividadeRespostaAluno;
         }
 
         public async Task SaveQuestaoResposta(List<QuestaoRespostaAluno> questaoRespostaAluno){
-            await _contexto.QuestaoRespostaAlunos.AddRangeAsync(questaoRespostaAluno);
+            await _contexto.QuestaoRespostaAluno.AddRangeAsync(questaoRespostaAluno);
             await _contexto.SaveChangesAsync();
         }
 
         public IEnumerable<AtividadeAluno> GetNumeroDeAtividadesFeitas(long idAluno)
         {
-            return _contexto.AtividadeAlunos.Where(at => at.IdPessoa == idAluno).ToList();
+            return _contexto.AtividadeAluno.Where(at => at.IdPessoa == idAluno).ToList();
         }
 
         public async Task AddAtividadeAluno(AtividadeAluno atividadeAluno)
         {
-            await _contexto.AtividadeAlunos.AddAsync(atividadeAluno);
+            await _contexto.AtividadeAluno.AddAsync(atividadeAluno);
             await _contexto.SaveChangesAsync();
         }
 
         public async Task UpdateAtividadeAluno(AtividadeAluno atividadeAluno)
         {
-            _contexto.AtividadeAlunos.Update(atividadeAluno);
+            _contexto.AtividadeAluno.Update(atividadeAluno);
             await _contexto.SaveChangesAsync();
         }
 
         public int GetAttemptsNumberFromId(long idAtividadeAluno){
-            AtividadeAluno atividadeAluno = _contexto.AtividadeAlunos.Where(a => a.IdAtividadeAluno == idAtividadeAluno).FirstOrDefault();
+            AtividadeAluno atividadeAluno = _contexto.AtividadeAluno.Where(a => a.IdAtividadeAluno == idAtividadeAluno).FirstOrDefault();
             if(atividadeAluno != null){
                 _contexto.Entry<AtividadeAluno>(atividadeAluno).State = Microsoft.EntityFrameworkCore.EntityState.Detached;
                 _contexto.SaveChanges();
